refactor: add generate HTML helper
